refactor : Product 내 price, ProductPrice 객체로 포장
